Depend only on "slf4j-api"
I think that your depending on several logging libraries is not a good thing. IMO, a clean approach would be to depend only on "slf4j-api" (and maybe one of its implementations in scope test) 

The change I'm proposing might break the build (I haven't tested it) but I just wanted to bring the topic to the table. Low level libraries such as yours, should only rely on "slf4j-api". The choosing of the implementation should be is up to the integrator.

If possible, you should also get rid of commons-logging. 

What do you think?